Cast to int to prevent warnings and type mismatches depending on what size_t actually is.
svn path=/icecast/trunk/libshout/; revision=11584

+++ examples/nonblocking.c
@@ -89,7 +89,8 @@ int main()
 				break;
 			}
 			if (shout_queuelen(shout) > 0)
-				printf("DEBUG: queue length: %d\n", shout_queuelen(shout));
+				printf("DEBUG: queue length: %d\n", 
+                                        (int)shout_queuelen(shout));
 
 			shout_sync(shout);
 		}
